Added Many Methods to the Record class:
Added a file to save all records to.
Added a method to load all records from the file.
Added a method to write all records to the file.
Added a method to add a new record.
Added a method to change between records.
Added a method to delete a record.
Added a private method to get the index of a record with a certain name.