Revision of Class Structure
Re-ordered the PianoMan class into two classes, and added action events
and references such that it should be possible to do all that needs to
be done, minus displaying the current rhythmic pattern. The code for the
PianoManCore class still needs to be written, and two of the four
instrumetns need files, and although the code "works", it is in need of
an overhaul once elvis finishes development on the AudioHandler class.